+++ app/Models/ServicoPmqaCampanhaPontoColeta.php
@@ -1,0 +1,23 @@
+<?php
+
+namespace App\Models;
+
+use Illuminate\Database\Eloquent\Factories\HasFactory;
+use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Relations\HasMany;
+
+class ServicoPmqaCampanhaPontoColeta extends Model
+{
+    use HasFactory;
+
+    protected $table = 'servico_pqma_campanha_ponto_coletas';
+    protected $guarded = ['id', 'created_at'];
+    protected $casts = [
+        'sem_coleta' => 'bool'
+    ];
+
+    public function arquivos(): HasMany
+    {
+        return $this->hasMany(ServicoPmqaCampanhaPontoColetaArquivo::class, 'coleta_id');
+    }
+}