+++ resources/js/Pages/Servico/PMQA/Navbar.vue
@@ -3,6 +3,8 @@ import NavDropdownLink from '@/Components/NavDropdownLink.vue';
 import NavDropdown from '@/Components/NavDropdown.vue';
 import Navbar from '@/Pages/Contrato/Contratada/Navbar.vue';
 import { IconLayoutDashboard } from '@tabler/icons-vue';
+import Breadcrumb from '@/Components/Breadcrumb.vue';
+import NavLink from '@/Components/NavLink.vue';
 
 const porps = defineProps({
   contrato: { type: Object },
@@ -12,6 +14,10 @@ const porps = defineProps({
 <template>
   <Navbar :contrato="contrato">
     <template #body>
+      <div class="mb-4">
+        <Breadcrumb
+          :links="[{ route: route('contratos.contratada.servicos.index', { contrato: contrato.id, servico: servico.id }), label: 'Serviços' }, { route: '#', label: servico?.tipo?.nome }]" />
+      </div>
       <div class="card card-body p-0 space-y-3">
         <header class="navbar-expand-md">
           <div class="collapse navbar-collapse" id="navbar-menu">
@@ -30,7 +36,16 @@ const porps = defineProps({
                         <NavDropdownLink route-name="contratos.contratada.servicos.pmqa.configuracao.parametro.index"
                           active-on-route-prefix="contratos.contratada.servicos.pmqa.configuracao.parametro*"
                           :route-param="{ contrato: contrato.id, servico: servico.id }" title="Parâmetros" />
+                        <NavDropdownLink
+                          route-name="contratos.contratada.servicos.pmqa.configuracao.vinculacao_ponto.index"
+                          active-on-route-prefix="contratos.contratada.servicos.pmqa.configuracao.vinculacao_ponto*"
+                          :route-param="{ contrato: contrato.id, servico: servico.id }" title="Vinculação de pontos" />
                       </NavDropdown>
+
+                      <NavLink route-name="contratos.contratada.servicos.pmqa.execucao.index"
+                        active-on-route-prefix="contratos.contratada.servicos.pmqa.execucao*"
+                        :param="{ contrato: contrato.id, servico: servico.id }" title="Execução"
+                        :icon="IconLayoutDashboard" />
                     </ul>
                   </div>
                 </div>